Move code from `decimal.pyi` to `_decimal.pyi`
